requirements.txt
Django>=4.2
pandas
matplotlib
seaborn
liac-arff
requests
gunicorn
whitenoise
